// apps/web/src/components/ui/Toast.d.ts
export declare const toast: {
    success: (msg: string, duration?: number) => void;
    error: (msg: string, duration?: number) => void;
    warning: (msg: string, duration?: number) => void;
    info: (msg: string, duration?: number) => void;
};
export declare function ToastContainer(): import("solid-js").JSX.Element;
//# sourceMappingURL=Toast.d.ts.map